add cart

+++ FreshFoodApp/app/src/main/java/com/example/freshfoodapp/Models/Account.java
@@ -6,24 +6,24 @@ import java.io.Serializable;
 
 public class Account  implements Serializable{
 
-    int id;
+    Long id;
 
     String name;
     String email;
     String avatar;
 
-    public Account(int id, String name, String email, String avatar) {
+    public Account(Long id, String name, String email, String avatar) {
         this.id = id;
         this.name = name;
         this.email = email;
         this.avatar = avatar;
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
